Move nested DTOs out of controller and HTTP client classes
Extracted all nested records from AdminAiProvidersController into
AiProviderDtos.cs and from AdminAiProvidersHttpClient into top-level
declarations in the same file. Removes mock SVG artefact.
