+++ code/FinanceManager.Components/HttpClients/AdminAiProvidersHttpClient.cs
@@ -2,30 +2,29 @@ using System.Net.Http.Json;
 
 namespace FinanceManager.Components.HttpClients;
 
+public sealed record AiProviderDto(
+    string ProviderName,
+    string BaseUrl,
+    bool HasApiKey,
+    int RequestTimeoutSeconds,
+    bool IsEnabled);
+
+public sealed record AiFallbackEntryDto(string ProviderName, string Model, int Order);
+
+public sealed record AiConfigurationResponse(
+    List<AiProviderDto> Providers,
+    List<AiFallbackEntryDto> FallbackEntries);
+
+public sealed record UpdateProviderRequest(
+    string BaseUrl,
+    string? ApiKey,
+    int RequestTimeoutSeconds,
+    bool IsEnabled);
+
+public sealed record UpdateFallbackRequest(List<AiFallbackEntryDto> Entries);
+
 public class AdminAiProvidersHttpClient(HttpClient httpClient)
 {
-    public sealed record AiProviderDto(
-        string ProviderName,
-        string BaseUrl,
-        bool HasApiKey,
-        int RequestTimeoutSeconds,
-        bool IsEnabled);
-
-    public sealed record AiFallbackEntryDto(string ProviderName, string Model, int Order);
-
-    public sealed record AiConfigurationResponse(
-        List<AiProviderDto> Providers,
-        List<AiFallbackEntryDto> FallbackEntries);
-
-    public sealed record UpdateProviderRequest(
-        string BaseUrl,
-        string? ApiKey,
-        int RequestTimeoutSeconds,
-        bool IsEnabled);
-
-    public sealed record UpdateFallbackRequest(
-        List<AiFallbackEntryDto> Entries);
-
     public async Task<AiConfigurationResponse?> GetConfigurationAsync(CancellationToken ct = default) =>
         await httpClient.GetFromJsonAsync<AiConfigurationResponse>(
             $"{httpClient.BaseAddress}api/admin/ai-providers", ct);
